+++ src/types.d.ts
@@ -30,35 +30,3 @@ export type CompletionItemWithScore = {
   item: CompletionItem | null;
   score: number;
 };
-
-type LinterConfigEnabler = {
-  enabled: boolean;
-};
-
-type LinterRubocopIgnores = {
-  ignored_cops: string[];
-} & LinterConfigEnabler;
-
-type LinterMetcher = {
-  file_types: 'partials' | 'all';
-  matchers: {
-    all: string;
-    partials: string;
-  };
-} & LinterConfigEnabler;
-
-/**
- * @see https://github.com/sds/haml-lint/blob/main/lib/haml_lint/linter/README.md
- */
-export type LinterConfig = {
-  RuboCop: LinterRubocopIgnores;
-  ClassesBeforeIds: LinterConfigEnabler;
-  FinalNewline: LinterConfigEnabler;
-  HtmlAttributes: LinterConfigEnabler;
-  LeadingCommentSpace: LinterConfigEnabler;
-  SpaceBeforeScript: LinterConfigEnabler;
-  StrictLocals: LinterMetcher;
-  TrailingEmptyLines: LinterConfigEnabler;
-  TrailingWhitespace: LinterConfigEnabler;
-  UnnecessaryStringOutput: LinterConfigEnabler;
-};